refactoring to add module for Word2vec related modules/classes
