+++ xn2v/w2v/__init__.py
@@ -1,0 +1,4 @@
+from .cbow_list_batcher import CBOWListBatcher
+
+
+__all__ = ["CBOWListBatcher"]
